Separated the routes and io from themain index.js
