+++ client/index.html
@@ -430,7 +430,7 @@
       </div>
     </div>
   </body>
-  <script src="jquerry.js"></script>
+  <script src="//code.jquery.com/jquery-latest.min.js"></script>
   <script src="client.js"></script>
   <script src="inventory.js"></script>
   <script src="canvas.js"></script>
